+++ include/input/macros/macro_pairing.h
@@ -4,6 +4,6 @@
 #include "input_shared_types.h"
 #include <stdint.h>
 
-void macro_pairing(uint32_t timestamp, button_data_s *buttons);
+void macro_pairing(uint64_t timestamp, button_data_s *buttons);
 
 #endif
